requirements.txt
django ==4.2.0
djangorestframework
pyyaml
requests
django-cors-headers
